Extend configurable service in SupportedVmService to be able to register it in service manager. Rename directory.
